Update airbyte-integrations/connectors/source-zendesk-sunshine/source_zendesk_sunshine/streams.py
backoff time int to float (btw real return type in headers is integer)

+++ airbyte-integrations/connectors/source-zendesk-sunshine/source_zendesk_sunshine/streams.py
@@ -48,7 +48,7 @@ class SunshineStream(HttpStream, ABC):
     def backoff_time(self, response: requests.Response) -> Optional[float]:
         delay_time = response.headers.get("Retry-After")
         if delay_time:
-            return int(delay_time)
+            return float(delay_time)
 
     def next_page_token(self, response: requests.Response) -> Optional[Mapping[str, Any]]:
         resp_json = response.json()
